+++ src/kio/schema/init_producer_id/v3/request.py
@@ -7,7 +7,6 @@ from typing import ClassVar
 
 from kio.schema.primitive import i16
 from kio.schema.primitive import i32
-from kio.schema.primitive import i64
 from kio.schema.types import ProducerId
 from kio.schema.types import TransactionalId
 
@@ -20,7 +19,7 @@ class InitProducerIdRequest:
     transaction_timeout_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The time in ms to wait before aborting idle transactions sent by this producer. This is only relevant if a TransactionalId has been defined."""
     producer_id: ProducerId = field(
-        metadata={"kafka_type": "int64"}, default=ProducerId(i64(-1))
+        metadata={"kafka_type": "int64"}, default=ProducerId(-1)
     )
     """The producer id. This is used to disambiguate requests if a transactional id is reused following its expiration."""
     producer_epoch: i16 = field(metadata={"kafka_type": "int16"}, default=i16(-1))
